requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.26.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
baukit @ git+https://github.com/davidbau/baukit@9d51abd51ebf29769aecc38c4cbef459b731a36e
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
bleach==6.1.0
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.4.0
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.16.0
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm @ file:///croot/comm_1709322850197/work
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.6
datasets==2.19.2
DateTime==5.5
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
decord==0.6.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
easyjailbreak==0.1.2
easyocr==1.7.1
einops==0.8.0
email_validator==2.1.1
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.6.3
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fschat==0.2.36
fsspec==2024.3.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.4
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
gradio==4.31.3
gradio_client==0.16.3
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.1
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.4.0
inquirerpy==0.3.4
iopath==0.1.10
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1694181358621/work
ipywidgets==8.1.2
isoduration==20.11.0
jcamp==1.2.2
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2 @ file:///croot/jinja2_1706733616596/work
jiter==0.4.0
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langsmith==0.1.59
lazy_loader==0.4
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml_collections==0.1.1
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1690561992265/work
nh3==0.2.17
ninja==1.11.1.1
nltk==3.8.1
notebook==7.2.0
notebook_shim==0.2.4
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.47.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
OpenOCR==0.1.0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.11.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pfzy==0.3.4
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow==10.3.0
platformdirs==4.2.2
playwright==1.47.0
portalocker==2.10.1
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
proto-plus==1.24.0
protobuf==4.25.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-bidi==0.4.2
python-dateutil @ file:///croot/python-dateutil_1715108784345/work
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
qtconsole==5.5.2
QtPy==2.4.1
rdkit==2024.3.5
referencing==0.35.1
regex==2024.5.15
replicate==0.33.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.4
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
selenium==4.21.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.3.1
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.37.2
svgwrite==1.4.3
sympy @ file:///croot/sympy_1701397643339/work
tenacity==8.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.7.0
timm==1.0.3
tinycss2==1.3.0
tokenizers==0.20.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.4
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.45.0
transformers-stream-generator==0.0.5
trio==0.25.1
trio-websocket==0.11.1
triton==2.3.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1715635851070/work
uvicorn==0.29.0
uvloop==0.19.0
wandb==0.17.0
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==1.13
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.10
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp @ file:///croot/zipp_1704206909481/work
zope.interface==7.0.3
